backend/websocket-service/requirements.txt
# WebSocket Server pour Djembé Bank
websockets>=12.0
redis[hiredis]>=5.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
